refactor pixels; don't make observation a package
