+++ ngmix/pixels/pixels_nb.py
