+++ ngmix/tests/test_observation.py
@@ -2,7 +2,7 @@ import numpy as np
 import pytest
 
 from ngmix.observation import Observation
-from ngmix.observation.pixels import make_pixels
+from ngmix.pixels import make_pixels
 from ngmix.jacobian import DiagonalJacobian
 from ngmix.gmix import GMix
